requirements.txt
aiohttp==3.13.3
apscheduler==3.11.2
fastapi==0.135.1
h5py==3.15.1
httpx==0.28.1
jsonschema==4.26.0
nest_asyncio==1.6.0
numpy==2.4.2
ollama==0.6.1
openai==2.24.0
postgrest==2.28.0
prompt_toolkit==3.0.52
psutil==7.2.2
pydantic==2.12.5
pyngrok==7.5.0
PyPDF2==3.0.1
python-dotenv==1.2.2
python_docx==1.1.0
redis==5.0.1
Requests==2.32.5
supabase==2.28.0
together==2.2.0
tqdm==4.67.1
uvicorn==0.41.0
